requirements.txt
Django==6.0.3
djangorestframework==3.16.1
drf-yasg==1.21.15
inflection==0.5.1
packaging==26.0
pytz==2026.1.post1
PyYAML==6.0.3
uritemplate==4.2.0
sqlparse==0.5.5
asgiref==3.11.1
tzdata==2025.3
